fix wild issue slack notification
