+++ src/webhooks/pubsub/slackNotifications.test.ts
@@ -7,7 +7,6 @@ import { db } from '@utils/db';
 
 import {
   constructSlackMessage,
-  getChannelsForIssue,
   getTriageSLOTimestamp,
   notifyProductOwnersForUntriagedIssues,
 } from './slackNotifications';
@@ -79,47 +78,7 @@ describe('Triage Notification Tests', function () {
       );
     });
   });
-  describe('getChannelsForIssue', () => {
-    it('will get channel info for repo without routing', () => {
-      expect(
-        getChannelsForIssue(
-          'test-ttt-simple',
-          'getsentry',
-          '',
-          moment('2022-12-12T17:00:00.000Z')
-        )
-      ).toEqual([{ channelId: 'C05A6BW303Z', isChannelInBusinessHours: true }]);
-    });
-    it('will get channel info for repo with routing', () => {
-      expect(
-        getChannelsForIssue(
-          'routing-repo',
-          'getsentry',
-          'Multi-Team',
-          moment('2022-12-14T00:00:00.000Z')
-        )
-      ).toEqual([
-        {
-          channelId: 'C05A6BW303Z',
-          isChannelInBusinessHours: true,
-        },
-        {
-          channelId: 'C05A6BW303B',
-          isChannelInBusinessHours: false,
-        },
-      ]);
-    });
-    it('will return team-ospo channel if inputs are invalid', () => {
-      expect(
-        getChannelsForIssue(
-          'garbage-repo',
-          'getsentry',
-          '',
-          moment('2022-12-12T17:00:00.000Z')
-        )
-      ).toEqual([{ channelId: 'C05A6BW303Y', isChannelInBusinessHours: true }]);
-    });
-  });
+
   describe('constructSlackMessage', function () {
     let boltPostMessageSpy;
     beforeEach(function () {
